requirements.txt
# Netskope -> ADX Function App (v2)
# Pin lower bounds, cap major versions to avoid breaking changes.

azure-functions>=1.21.3,<2.0.0
azure-kusto-data>=4.4.1,<5.0.0
azure-kusto-ingest>=4.4.1,<5.0.0
azure-identity>=1.19.0,<2.0.0
requests>=2.32.0,<3.0.0
urllib3>=2.0.0,<3.0.0
